scripts/requirement.txt
binance-connector-python==0.6.0
mysql-connector-python==8.0.32
pandas==1.5.3
pycryptodome>=3.15.0
psutil==5.9.4
termcolor==2.2.0
websocket-client==1.5.1
fastapi==0.95.0